add roles time.difference, time.timeout and time.interval
Currently no roles exist to specify a timeout values for i.e. for network operations.
And the same applies for states controlling an intervall for recurrent operations.
time.difference would be useful to specify the difference of 2 times, ie. the duration of last operation or the deviation from a planned timestamp

I suggest to specify the values as ms as ms is the technical unit für javascript timers 

@Apollon77 
@GermanBluefox 
@Garfonso 

Solves https://github.com/ioBroker/ioBroker.type-detector/issues/36