Fix line breaks and add dots at end of sentences
Parts of some sentences were broken up into separate paragraphs when rendered as Markdown because an end-of-line was added in the middle of the sentence and the next line was not indented.

Rather than add indentation on the next lines, I've opted to remove the end-of-lines. This seems more reasonable in a time where different devices have different screen width and auto-line-wrap is implemented in almost every editor.