Update state roles documentation for level.effect
Added description for 'level.effect' and fixed formatting for 'level.color.cie'.

Some lights offer a list of effects that the device itself can do, like flashing or fading or whatever (some can do quite complex stuff). Adapters today implement that as state with `common.type=string` and a list of possible effects in `common.states`. 
I'd like to add support for an effect state in type-detector for most of the light types (and after that add it to lovelace, because the frontend seems to support that already).

`common.type=string` is not really correct for `level`. But there are already some exceptions to that rule, mostly used for lights already.